requirements.txt
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==22.1
factory-boy==3.0.1
black==22.10.0
